+++ runtime_contract.go
@@ -3,7 +3,7 @@ package openlinker
 const (
 	RuntimeProtocolVersion = 2
 	RuntimeContractID      = "openlinker.runtime.v2"
-	RuntimeContractDigest  = "857598f6e8f07d87d1f7240e34d98f0911bf23e5204a865d282a6bcb7f52865f"
+	RuntimeContractDigest  = "052ed16553eeb896bc7a88dabd1ada77466a4db0c87b55c997c6b91ab72a72de"
 )
 
 var runtimeRequiredFeatures = [...]string{